+++ server/app/providers/ocblacktop.py
@@ -21,7 +21,7 @@ class OcBlackTopProvider(Provider):
         headers = {
             "Accept": "application/json",
             "x-api-key": settings.ocblacktop_api_key,
-            "User-Agent": "Motorsports-Events-Server/2.5.2",
+            "User-Agent": "Motorsports-Events-Server/2.6.0",
         }
 
         results: list[NormalizedEvent] = []